+++ R/micro2intData.r
@@ -2,7 +2,7 @@
 #'
 #' Aggregates microdata from a data frame into interval-valued data using various criteria and latent distribution settings.
 #'
-#' @param MicDtDF A data frame containing the microdata. All columns should be numeric.
+#' @param microdata A data frame containing the microdata. All columns should be numeric.
 #' @param agrby A factor used to specify the grouping of the microdata for aggregation.
 #' @param agrcrt A string or numeric vector of length 2 specifying the aggregation criterion. The default is \code{"minmax"}, which takes the minimum and maximum values for each variable. If a numeric vector is provided, it should specify the lower and upper percentiles for aggregation (e.g., \code{c(0.05, 0.95)}).
 #' @param LatentParam (Optional) A list with the parameters of the latent variables. 
@@ -50,7 +50,7 @@
 #'                             LatentCase = "General")
 #' 
 #' @export
-micro2intData <- function(MicDtDF,
+micro2intData <- function(microdata,
                           agrby,
                           agrcrt="minmax",
                           LatentParam=NULL,
@@ -60,57 +60,57 @@ micro2intData <- function(MicDtDF,
                           BetaParam.a=1,
                           BetaParam.b=1,
                           estimate.DistParam=FALSE){
-  mcall <- match.call()$MicDtDF
+  mcall <- match.call()$microdata
   if (length(mcall) > 1) mcall <- "microdata"
-  if (!(is.data.frame(MicDtDF))) stop("First argument of AgMicroData must be a data frame.\n")
-  if (!is.data.frame(MicDtDF)) MicDtDF <- as.data.frame(MicDtDF)
-  if (any(!sapply(seq_len(ncol(MicDtDF)),function(ind) is.numeric(MicDtDF[,ind])))){  
+  if (!(is.data.frame(microdata))) stop("First argument of AgMicroData must be a data frame.\n")
+  if (!is.data.frame(microdata)) microdata <- as.data.frame(microdata)
+  if (any(!sapply(seq_len(ncol(microdata)),function(ind) is.numeric(microdata[,ind])))){  
     stop(paste("Some of the columns of the",mcall,"data frame have non-numeric variables.\n"))
   }
   
-  unvalidobs <- which(apply(MicDtDF,1,function(v) all(!is.finite(v))))
+  unvalidobs <- which(apply(microdata,1,function(v) all(!is.finite(v))))
   nunvalid <- length(unvalidobs) 
   if (nunvalid>0) {
-    MicDtDF <- MicDtDF[-unvalidobs,]
+    microdata <- microdata[-unvalidobs,]
     agrby <- agrby[-unvalidobs]
     string2 <- paste("rows of the",mcall,"data frame were dropped because they only included non-valid (non finite or missing values) observations.\n")
-    if (nunvalid<=10) warning(paste("The",paste(row.names(MicDtDF)[unvalidobs],collapse=" "),string2))
+    if (nunvalid<=10) warning(paste("The",paste(row.names(microdata)[unvalidobs],collapse=" "),string2))
     else warning(paste(nunvalid,string2,collapse=" "))
   }
 
   if (!is.factor(agrby)) stop("Argument agrby is not a factor\n")
-  globaln <- nrow(MicDtDF)
-  if (length(agrby)!=globaln) stop("Size of the agrby argument does not agree with the number of rows in the MicDtDF data frame.\n") 
+  globaln <- nrow(microdata)
+  if (length(agrby)!=globaln) stop("Size of the agrby argument does not agree with the number of rows in the microdata data frame.\n") 
   if ( agrcrt[1]!="minmax" && (!is.numeric(agrcrt) || length(agrcrt)!=2 || agrcrt[1]>=agrcrt[2] || agrcrt[1]<0. || agrcrt[2]>1.) )
     stop(paste("Wrong value for the agrcrt argument\n( it should be either the string minmax or a two-dim vector",
                "\nof a prob. value for the lower percentile, followed by the prob. value for the upper percentile - \nex:c(0.05,0.95) ).\n")) 
   
   # Trim microdata, if applicable
   if (agrcrt[1] != "minmax") {
-    MicDtDF_trim <- MicDtDF
-    split_idx <- split(seq_len(nrow(MicDtDF)), agrby)
+    microdata_trim <- microdata
+    split_idx <- split(seq_len(nrow(microdata)), agrby)
     for (grp in names(split_idx)) {
       idx <- split_idx[[grp]]
-      for (col in seq_len(ncol(MicDtDF))) {
-        x <- MicDtDF[idx, col]
+      for (col in seq_len(ncol(microdata))) {
+        x <- microdata[idx, col]
         if (!all(is.na(x))) {
           q <- quantile(x, probs = agrcrt, na.rm = TRUE)
           sel <- which(x < q[1] | x > q[2])
-          if (length(sel) > 0) MicDtDF_trim[idx[sel], col] <- NA
+          if (length(sel) > 0) microdata_trim[idx[sel], col] <- NA
         }
       }
     }
-    MicDtDF <- MicDtDF_trim
+    microdata <- microdata_trim
   }
 
   if (length(unique(agrby))!=length(levels(agrby)))  agrby <- factor(agrby)
   grplvls <- levels(agrby)
-  nvar <- ncol(MicDtDF)
+  nvar <- ncol(microdata)
 
   # logical vector: TRUE = group is valid; FALSE = group has at least one variable all NA
   keep_group <- sapply(grplvls, function(g) {
     rind <- which(agrby == g)
-    all_na_in_any_var <- any(sapply(1:nvar, function(c) all(is.na(MicDtDF[rind, c]))))
+    all_na_in_any_var <- any(sapply(1:nvar, function(c) all(is.na(microdata[rind, c]))))
     !all_na_in_any_var  # keep if FALSE
   })
 
@@ -127,28 +127,28 @@ micro2intData <- function(MicDtDF,
 
     # subset data and grouping factor
     valid_idx <- agrby %in% valid_grplvls
-    MicDtDF <- MicDtDF[valid_idx, , drop = FALSE]
+    microdata <- microdata[valid_idx, , drop = FALSE]
     agrby   <- droplevels(agrby[valid_idx])
     grplvls <- valid_grplvls
-    nvar <- ncol(MicDtDF)
+    nvar <- ncol(microdata)
   }
 
   ngrps <- length(grplvls)
   bndsDF <- as.data.frame(matrix(NA_real_, nrow = ngrps, ncol = 2 * nvar))
-  NbMicroUnits <- integer(ngrps)
+  NMicro <- integer(ngrps)
   for (r in 1:ngrps) { 
     grp <- grplvls[r]
     rind <- which(agrby==grp)
-    NbMicroUnits[r] <- length(rind)
+    NMicro[r] <- length(rind)
     for (c in 1:nvar) {
-      bndsDF[r,c] <- min(MicDtDF[rind,c], na.rm = TRUE)
-      bndsDF[r,nvar+c] <- max(MicDtDF[rind,c], na.rm = TRUE)
+      bndsDF[r,c] <- min(microdata[rind,c], na.rm = TRUE)
+      bndsDF[r,nvar+c] <- max(microdata[rind,c], na.rm = TRUE)
     }
   }
   rownames(bndsDF) <- grplvls
   
-  Umicro <- get_latent_var(MicDtDF,bndsDF,agrby,agrlevels=grplvls,Seq="AllLb_AllUb")
-  res <- intData(bndsDF,Seq="AllLb_AllUb",LatentParam,LatentCase,LatentDist,TriangParam,BetaParam.a,BetaParam.b,Umicro,estimate.DistParam,VarNames=names(MicDtDF),ObsNames=grplvls)
+  Umicro <- get_latent_var(microdata,bndsDF,agrby,agrlevels=grplvls,Seq="AllLb_AllUb")
+  res <- intData(bndsDF,Seq="AllLb_AllUb",LatentParam,LatentCase,LatentDist,TriangParam,BetaParam.a,BetaParam.b,Umicro,estimate.DistParam,VarNames=names(microdata),ObsNames=grplvls)
   DegInT <- which(apply(res@Ranges,1,function(v) any(v==0)))
   nDegInT <- length(DegInT)
   if (nDegInT>0) {
@@ -169,10 +169,10 @@ micro2intData <- function(MicDtDF,
     }
     warning(wmsg)
     res <- res[-DegInT,]
-    res@NbMicroUnits <- NbMicroUnits[-DegInT]
+    res@NMicro <- NMicro[-DegInT]
   } else {
-    res@NbMicroUnits <- NbMicroUnits
+    res@NMicro <- NMicro
   }  
-  names(res@NbMicroUnits) <- res@ObsNames
+  names(res@NMicro) <- res@ObsNames
   res
 }