+++ NEWS.md
@@ -31,6 +31,7 @@
 
 * Introduced functionality for explainable outlier detection using Shapley values.
 * Added functions to compute feature contributions and interaction effects: `int_Shapley()`, `int_Shapley_decomp()`, and `int_Shapley_interaction()`. 
-* Added visualization functions for Shapley values and Shapley interaction indices: `barplot_int_Shapley_decomp()`, `barplot_int_Shapley()`, `beeswarm_int_Shapley()`, `plot_int_Shapley_inter()`, and `radarplot_int_Shapley()`.
+* Added visualization functions for Shapley values and Shapley interaction indices: `plot_bar_int_Shapley_decomp()`, `plot_bar_int_Shapley()`, `plot_beeswarm_int_Shapley()`, `plot_int_Shapley_inter()`, and `plot_radar_int_Shapley()`.
+* Functions `SYMB.biplot` and `SYMB.pairs.panels` renamed to `plot_scatter_int`and `plot_pairs_int`, respectively.
 * Improvements to documentation and usability.
 * Added a unit test suite.